Create Query Interface
* Created query interface for increased modularity
* Improved import flow
* Standardized backend function names
